+++ Form/ContactFormType.php
@@ -11,6 +11,7 @@ namespace c975L\ContactFormBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
@@ -18,7 +19,6 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class ContactFormType extends AbstractType
 {
-
     //Builds the form
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
@@ -52,8 +52,16 @@ class ContactFormType extends AbstractType
                     'placeholder' => 'placeholder.message',
                 )))
         ;
+        if ($options['receiveCopy'] === true) {
+            $builder
+                ->add('receiveCopy', CheckboxType::class, array(
+                    'label' => 'label.receive_copy',
+                    'required' => false,
+                    'data' => false,
+                    ))
+            ;
+        }
     }
-
 
     public function configureOptions(OptionsResolver $resolver)
     {
@@ -62,5 +70,7 @@ class ContactFormType extends AbstractType
             'intention'  => 'contactForm',
             'translation_domain' => 'contactForm',
         ));
+
+        $resolver->setRequired('receiveCopy');
     }
 }